requirements.txt
peewee
pymysql
cryptography
grpcio
grpcio-tools
loguru
requests
python-consul